Reorganize the code into logical packages and add structure to the flat files
